+++ src/main/java/me/kqlqk/behealthy/condition_service/service/KcalsInfoService.java
@@ -10,5 +10,5 @@ import org.springframework.stereotype.Service;
 public interface KcalsInfoService {
     KcalsInfo getById(long id);
 
-    KcalsInfo generateDailyKcals(Gender gender, byte age, short height, short weight, Intensity intensity, Goal goal);
+    KcalsInfo generateDailyKcals(Gender gender, int age, int height, int weight, Intensity intensity, Goal goal);
 }
